moving createJwtIds method to tokenIdService for clear seperation of concern
