+++ E-Stores-API/src/main/java/com/devb/estores/service/TokenIdService.java
@@ -4,6 +4,8 @@ import com.devb.estores.enums.TokenType;
 
 public interface TokenIdService {
 
+    String generateJwtId(String username, String deviceId, TokenType tokenType);
+
     String getJti(String username, String deviceId, TokenType tokenType);
 
     void deleteJti(String username, String deviceId);
