+++ E-Stores-API/src/main/java/com/devb/estores/serviceimpl/AuthServiceImpl.java
@@ -8,9 +8,7 @@ import com.devb.estores.dto.OtpModel;
 import com.devb.estores.enums.TokenType;
 import com.devb.estores.enums.UserRole;
 import com.devb.estores.exceptions.*;
-import com.devb.estores.model.TokenIdentification;
 import com.devb.estores.model.User;
-import com.devb.estores.repository.TokenIdentificationRepo;
 import com.devb.estores.repository.UserRepo;
 import com.devb.estores.requestdto.AuthRequest;
 import com.devb.estores.requestdto.UserRequest;
@@ -36,7 +34,6 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDateTime;
 import java.util.*;
 
 @Slf4j
@@ -53,7 +50,6 @@ public class AuthServiceImpl implements AuthService {
     private final Random random;
     private final AppEnv appEnv;
     private final CacheService cacheService;
-    private final TokenIdentificationRepo tokenIdRepo;
     private final TokenIdService tokenIdService;
 
     public static final String FAILED_REFRESH = "Failed to refresh login";
@@ -169,7 +165,7 @@ public class AuthServiceImpl implements AuthService {
 
         long expiration = (tokenType.equals(TokenType.ACCESS)) ? appEnv.getJwt().getAccessExpirationSeconds(): appEnv.getJwt().getRefreshExpirationSeconds();
 
-        String jti = generateJwtId(username, deviceId, tokenType);
+        String jti = tokenIdService.generateJwtId(username, deviceId, tokenType);
 
         TokenPayload tokenPayload = TokenPayload.create()
                 .setSubject(username)
@@ -200,44 +196,6 @@ public class AuthServiceImpl implements AuthService {
             return secChUa.substring(start, end);
         } else
             return null;
-    }
-
-    /**
-     * Generates a new JTI and caches it to respective cache's,
-     * the JTI is cached with the key created by combining username and deviceId.
-     * Like, username + "." + deviceId
-     *
-     *  @param username username of the account.
-     *  @param deviceId deviceId i.e, created for the client device identification.
-     *  @param tokenType the token type to which the JTI is to be created.
-     *  */
-    private String generateJwtId(String username, String deviceId, TokenType tokenType) {
-        String jti = UUID.randomUUID().toString();
-        String key = username + "." + deviceId;
-
-        // Caching jti for token identification in future
-        switch (tokenType) {
-            case ACCESS -> {
-                this.saveJtiForBackup(username, deviceId, tokenType, jti, LocalDateTime.now().plusSeconds(appEnv.getJwt().getAccessExpirationSeconds()));
-                cacheService.doEntry(CacheName.ACCESS_TOKEN_CACHE, key, jti);
-            }
-            case REFRESH -> {
-                this.saveJtiForBackup(username, deviceId, tokenType, jti, LocalDateTime.now().plusSeconds(appEnv.getJwt().getRefreshExpirationSeconds()));
-                cacheService.doEntry(CacheName.REFRESH_TOKEN_CACHE, key, jti);
-            }
-        }
-
-        return jti;
-    }
-
-    private void saveJtiForBackup(String username, String deviceId, TokenType tokenType, String jti, LocalDateTime expiration) {
-        tokenIdRepo.save(TokenIdentification.builder()
-                .username(username)
-                .deviceId(deviceId)
-                .tokenType(tokenType)
-                .jti(jti)
-                .expiration(expiration)
-                .build());
     }
 
     @Override
